+++ utils/handlers.py
@@ -10,11 +10,38 @@ from telegram.ext import (
     Filters,
     MessageHandler,
 )
+from utils.bot_status import set_bot_status
 from utils.keyboard import create_markup, get_reply_keyboard_markup
 from utils.message_strings import *
 from utils.jokes import getJoke
 from utils.states import *
 from utils import mongo
+
+from bson.json_util import dumps
+
+
+def submit_details(update, context):
+    update.message.reply_text(
+        text=PROCEED_MESSAGE, parse_mode=telegram.ParseMode.MARKDOWN
+    )
+    update.message.reply_text(
+        text='Please click on "Submit Details" to proceed',
+        parse_mode=telegram.ParseMode.MARKDOWN,
+        reply_markup=create_markup([["Submit Details"], ["Cancel"]]),
+    )
+    return FOLLOW_TELEGRAM
+
+
+def follow_twitter(update, context):
+    update.message.reply_text(
+        text=FOLLOW_TWITTER_TEXT, parse_mode=telegram.ParseMode.MARKDOWN
+    )
+    update.message.reply_text(
+        text="Type in *your Twitter username* to proceed",
+        parse_mode=telegram.ParseMode.MARKDOWN,
+        reply_markup=create_markup([["Cancel"]]),
+    )
+    return SUBMIT_ADDRESS
 
 
 def maxNumberReached(update, context):
@@ -148,3 +175,54 @@ def sureWantTo(update, context):
             reply_markup=get_reply_keyboard_markup(),
         )
         return LOOP
+
+
+# Admin commands
+def getList(update, context):
+    user = update.message.from_user
+    if user.username != ADMIN_USERNAME:
+        return
+    list = mongo.users.find({})
+
+    with open("users.json", "w") as file:
+        file.write("[")
+        for document in list:
+            file.write(dumps(document))
+            file.write(",")
+        file.write("]")
+    with open("users.json", "r") as file:
+        update.message.reply_document(document=file, filename="list.json")
+
+
+def getStats(update, context):
+    user = update.message.from_user
+    if user.username != ADMIN_USERNAME:
+        return
+    list = mongo.users.find({})
+    refes = mongo.users.find({"ref": {"$ne": False}}).count()
+    reply = f"""
+Currently there are *{list.count()} users* joined the airdrop!
+Currently there are *{refes} users* joined by referrals
+A total of *{"{:,.2f}".format(float(AIRDROP_AMOUNT.replace(",",""))*list.count())} {COIN_SYMBOL}* will be distributed as participation rewards
+A total of *{"{:,.2f}".format(REFERRAL_REWARD*refes)} {COIN_SYMBOL}* referral rewards will be distributed
+"""
+    update.message.reply_text(reply, parse_mode=telegram.ParseMode.MARKDOWN)
+
+
+def setStatus(update, context):
+    user = update.message.from_user
+    if user.username != ADMIN_USERNAME:
+        return
+    arg = context.args[0]
+    if arg == "stop":
+        set_bot_status("STOPPED")
+        update.message.reply_text("Airdrop stopped")
+    if arg == "pause":
+        set_bot_status("PAUSED")
+        update.message.reply_text("Airdrop paused")
+    if arg == "start":
+        set_bot_status("ON")
+        update.message.reply_text("Airdrop started")
+
+
+# ----------------
